web/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
openai==1.54.5
python-dotenv==1.0.1
